feat: share photo metadata flyout
